+++ order-service/src/main/java/com/node5/orderservice/config/OpenFeignConfig.java
@@ -1,9 +1,9 @@
 package com.node5.orderservice.config;
 
-import com.node5.orderservice.global.exception.openfeign.client.WalletClient;
-import com.node5.orderservice.global.exception.openfeign.client.CatalogClient;
-import com.node5.orderservice.global.exception.openfeign.client.SettlementClient;
-import com.node5.orderservice.global.exception.openfeign.client.MemberClient;
+import com.node5.orderservice.global.openfeign.client.WalletClient;
+import com.node5.orderservice.global.openfeign.client.CatalogClient;
+import com.node5.orderservice.global.openfeign.client.SettlementClient;
+import com.node5.orderservice.global.openfeign.client.MemberClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
