+++ order-service/src/main/java/com/node5/orderservice/global/openfeign/client/dto/StockHoldBatchRequest.java
@@ -1,8 +1,4 @@
-package com.node5.orderservice.global.exception.openfeign.client.dto;
-
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
+package com.node5.orderservice.global.openfeign.client.dto;
 
 import java.util.List;
 import java.util.UUID;
